Change:Relocated files under main method folder.Reason for white color page error.
