+++ src/main/java/com/example/_2ch/app/entity/Role.java
@@ -1,4 +1,4 @@
-package com.example._2ch.entity;
+package com.example._2ch.app.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
