+++ src/main/java/com/example/_2ch/app/repository/RoleRepository.java
@@ -1,6 +1,6 @@
-package com.example._2ch.repository;
+package com.example._2ch.app.repository;
 
-import com.example._2ch.entity.Role;
+import com.example._2ch.app.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoleRepository extends JpaRepository<Role, Long>{
